feat(waveform): implement 9 waveforms per NNC6521 spec
- Add burst_pulse_64 and deep_sculpt_pulse_128 waveform data arrays
- Rewrite waveform_apply() with proper gen methods per spec:
  #1 Power Smooth: Preloaded PULSE (50Hz, 300us)
  #2 Burst Train: Customized SPI burst_pulse_64 (50Hz)
  #3 Gentle Smooth: Preloaded PULSE (35Hz, 300us)
  #4 Deep Sculpt: Customized SPI deep_sculpt_pulse_128 (4kHz carrier)
  #5 Soft Sculpt: Customized SPI sine 128 (40Hz)
  #6 Circulation Sculpt: AM sine 64 (4kHz carrier, 10Hz envelope)
  #7 Smooth & Firm: Preloaded TRIANGLE (100Hz, 400us)
  #8 Lymphatic Drainage: Customized SPI sine 64 (5Hz, PCLK/16)
  #9 Soothing Ending: Customized SPI sine 64 (10Hz, PCLK/8)
- Fix uint16_t overflow for waveforms #8/#9 via dynamic PCLK divider
- Add waveform info print in spec format
- Update protocol_update_current_output() to use waveform_apply()

+++ workItem-djm-v10/applications/macBSP/Inc/nnc6521_reg.h
@@ -522,11 +522,11 @@ typedef union {
 } WG_DrvSilentClkReg_t;
 
 typedef union {
-    uint16_t value;
+    uint32_t value;
 } WG_DrvHalfWaveClkPntReg_t;
 
 typedef union {
-    uint16_t value;
+    uint32_t value;
 } WG_DrvNegHalfWaveClkPntReg_t;
 
 typedef union {
